requirements.txt
pandas
finta
python-binance
